Create ContributionActivity
For the ESOP tracking, make a report that shows how much was contributed each month. Distinguish if it was to a forfeiture, suspense, or general/participant/TM account.

Make a column for each of the following. Cash Positive, cash negative, Shares positive, shares negative.

This should be only for the contributions.